Move exercises to base package (since "lesson" is fairly ambiguous)
