+++ intro-to-java/src/main/java/eca/exercises/BACA/BasicCalculator.java
@@ -1,0 +1,19 @@
+package eca.exercises.BACA;
+
+public class BasicCalculator {
+
+    public static void calculate() {
+        // Use UserInputUtil to take in Integers a and b, and String operator
+
+
+        // Determine which math operator to do based on the operator String
+
+        // Then perform that operation on a and b
+        // Print the math operation and the result e.g. 1 + 2 = 3
+
+    }
+
+    public static void main(String[] args) {
+        calculate();
+    }
+}
