+++ intro-to-java/src/main/java/eca/exercises/GWNT/GuessWhatNumberImThinking.java
@@ -1,0 +1,23 @@
+package eca.exercises.GWNT;
+
+public class GuessWhatNumberImThinking {
+
+    static String YOU_WIN = "You Win!";
+    static String YOU_LOSE = "You LOSE!";
+
+    public static void guessWhatNumber() {
+        // Get a random number
+
+        // Let the user enter a number
+
+        // If the user's number matches the random number, they win!
+
+        // Print the result
+        System.out.println(YOU_LOSE);
+    }
+
+
+    public static void main(String[] args) {
+        guessWhatNumber();
+    }
+}
