+++ intro-to-java/src/main/java/eca/exercises/HINH/HelloInsertNameHere.java
@@ -1,0 +1,13 @@
+package eca.exercises.HINH;
+
+public class HelloInsertNameHere {
+
+    public static void main(String[] args) {
+        // Get the user's name using UserInputUtil
+
+
+        // Construct and print the Hello String
+
+
+    }
+}
