+++ intro-to-java/src/main/java/eca/exercises/MAOP/MathOperations.java
@@ -1,0 +1,20 @@
+package eca.exercises.MAOP;
+
+public class MathOperations {
+
+    public static void main(String[] args) {
+        // Create 2 Integer variables a and b
+
+
+        /* Print the result of:
+            - Addition
+            - Subtraction
+            - Multiplication
+            - Division
+            - Modulus
+           between a and b.
+         */
+
+
+    }
+}
