+++ intro-to-java/src/main/java/eca/exercises/MCMP/MathComparisons.java
@@ -1,0 +1,19 @@
+package eca.exercises.MCMP;
+
+public class MathComparisons {
+
+    public static void main(String[] args) {
+        // Create 2 Integer variables a and b
+
+
+        /* Print the result of:
+            - Equal to
+            - Greater than
+            - Greater than or equal
+            - Less than
+            - Less than or equal
+           between a and b.
+         */
+
+    }
+}
