+++ intro-to-java/src/main/java/eca/exercises/OOEV/OddOrEven.java
@@ -1,0 +1,13 @@
+package eca.exercises.OOEV;
+
+public class OddOrEven {
+
+    public static Boolean isOdd(Integer number) {
+        // Replace false with the result of your solution
+        return false;
+    }
+
+    public static void main(String[] args) {
+        System.out.println(isOdd(4));
+    }
+}
