+++ intro-to-java/src/main/java/eca/exercises/PYCA/PythagorasCalculation.java
@@ -1,0 +1,19 @@
+package eca.exercises.PYCA;
+
+public class PythagorasCalculation {
+
+    public static Double calculatePythagoras(Integer sideA, Integer sideB) {
+        // Using sideA and sideB, calculate the length of sideC of the triangle using Pythagoras' theorem
+
+
+        // Print your message
+
+
+        // Return the value of sideC
+        return 0.0;
+    }
+
+    public static void main(String[] args) {
+        Double length = calculatePythagoras(1, 2);
+    }
+}
