+++ intro-to-java/src/main/java/eca/exercises/VABO/VariablesAndBasicOperators.java
@@ -1,0 +1,16 @@
+package eca.exercises.VABO;
+
+public class VariablesAndBasicOperators {
+
+    public static void main(String[] args) {
+        // Define your String, Integer, and Boolean
+
+
+        // Print them using System.out.println
+
+
+        // Print them together, forming a sentence (that may or may not make sense!)
+
+
+    }
+}
