--- intro-to-java/src/main/java/eca/lessons/lesson2/exercises/BACA/BasicCalculator.java
@@ -1,21 +1,0 @@
-package eca.lessons.lesson2.exercises.BACA;
-
-import eca.util.UserInputUtil;
-
-public class BasicCalculator {
-
-    public static void calculate() {
-        // Use UserInputUtil to take in Integers a and b, and String operator
-
-
-        // Determine which math operator to do based on the operator String
-
-        // Then perform that operation on a and b
-        // Print the math operation and the result e.g. 1 + 2 = 3
-
-    }
-
-    public static void main(String[] args) {
-        calculate();
-    }
-}
