--- intro-to-java/src/main/java/eca/lessons/lesson2/exercises/GWNT/GuessWhatNumberImThinking.java
@@ -1,26 +1,0 @@
-package eca.lessons.lesson2.exercises.GWNT;
-
-import eca.util.UserInputUtil;
-import java.util.Random;
-
-public class GuessWhatNumberImThinking {
-
-    static String YOU_WIN = "You Win!";
-    static String YOU_LOSE = "You LOSE!";
-
-    public static void guessWhatNumber() {
-        // Get a random number
-
-        // Let the user enter a number
-
-        // If the user's number matches the random number, they win!
-
-        // Print the result
-        System.out.println(YOU_LOSE);
-    }
-
-
-    public static void main(String[] args) {
-        guessWhatNumber();
-    }
-}
