--- intro-to-java/src/main/java/eca/lessons/lesson2/exercises/HINH/HelloInsertNameHere.java
@@ -1,15 +1,0 @@
-package eca.lessons.lesson2.exercises.HINH;
-
-import eca.util.UserInputUtil;
-
-public class HelloInsertNameHere {
-
-    public static void main(String[] args) {
-        // Get the user's name using UserInputUtil
-
-
-        // Construct and print the Hello String
-
-
-    }
-}
