--- intro-to-java/src/main/java/eca/lessons/lesson2/exercises/MAOP/answers/AnswersMAOP.java
@@ -1,24 +1,0 @@
-package eca.lessons.lesson2.exercises.MAOP.answers;
-
-public class AnswersMAOP {
-
-    public static void main(String[] args) {
-        // Create 2 Integer variables a and b
-        Integer a = 10;
-        Integer b = 2;
-
-        /* Print the result of:
-            - Addition
-            - Subtraction
-            - Multiplication
-            - Division
-            - Modulus
-           between a and b.
-         */
-        System.out.println(a + b); // Addition
-        System.out.println(a - b); // Subtraction
-        System.out.println(a * b); // Multiplication
-        System.out.println(a / b); // Division
-        System.out.println(a % b); // Modulus (The remainder of a / b)
-    }
-}
