--- intro-to-java/src/main/java/eca/lessons/lesson2/exercises/OOEV/OddOrEven.java
@@ -1,13 +1,0 @@
-package eca.lessons.lesson2.exercises.OOEV;
-
-public class OddOrEven {
-
-    public static Boolean isOdd(Integer number) {
-        // Replace false with the result of your solution
-        return false;
-    }
-
-    public static void main(String[] args) {
-        System.out.println(isOdd(4));
-    }
-}
