--- intro-to-java/src/main/java/eca/lessons/lesson2/exercises/PYCA/PythagorasCalculation.java
@@ -1,21 +1,0 @@
-package eca.lessons.lesson2.exercises.PYCA;
-
-import java.math.*;
-
-public class PythagorasCalculation {
-
-    public static Double calculatePythagoras(Integer sideA, Integer sideB) {
-        // Using sideA and sideB, calculate the length of sideC of the triangle using Pythagoras' theorem
-
-
-        // Print your message
-
-
-        // Return the value of sideC
-        return 0.0;
-    }
-
-    public static void main(String[] args) {
-        Double length = calculatePythagoras(1, 2);
-    }
-}
