--- intro-to-java/src/main/java/eca/lessons/lesson2/exercises/SETE/SumEqualToExpected.java
@@ -1,13 +1,0 @@
-package eca.lessons.lesson2.exercises.SETE;
-
-public class SumEqualToExpected {
-
-    public static Boolean sumEqualToExpected(Integer a, Integer b, Integer expected) {
-        // Replace false with the result of your solution
-        return false;
-    }
-
-    public static void main(String[] args) {
-
-    }
-}
