--- intro-to-java/src/main/java/eca/lessons/lesson2/exercises/VABO/VariablesAndBasicOperators.java
@@ -1,16 +1,0 @@
-package eca.lessons.lesson2.exercises.VABO;
-
-public class VariablesAndBasicOperators {
-
-    public static void main(String[] args) {
-        // Define your String, Integer, and Boolean
-
-
-        // Print them using System.out.println
-
-
-        // Print them together, forming a sentence (that may or may not make sense!)
-
-
-    }
-}
