--- intro-to-java/src/test/java/eca/lessons/lesson2/exercises/SETE/SumEqualToExpectedTest.java
@@ -1,30 +1,0 @@
-package eca.lessons.lesson2.exercises.SETE;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class SumEqualToExpectedTest {
-
-    @Test
-    public void shouldReturnTrueForMatchingResult() {
-        Integer a = 5;
-        Integer b = 2;
-        Integer expected = 7;
-
-        Boolean result = SumEqualToExpected.sumEqualToExpected(a, b, expected);
-
-        assertTrue(result);
-    }
-
-    @Test
-    public void shouldReturnFalseForNonMatchingResult() {
-        Integer a = 5;
-        Integer b = 2;
-        Integer expected = 10;
-
-        Boolean result = SumEqualToExpected.sumEqualToExpected(a, b, expected);
-
-        assertFalse(result);
-    }
-}
